requirements.txt
python-telegram-bot==20.7
openai==1.3.7
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
flask==3.0.0
jinja2==3.1.2